pyproject.toml
[project]
name = "sokrates-mcp"
version = "0.1.6"
description = "A templated MCP server for demonstration and quick start."
requires-python = ">=3.10"
dependencies = [
    "fastmcp",
    "sokrates",
#    "sokrates @ file:///Users/dude/dev/os_projects/sokrates",
    "pydantic",
    "PyYAML"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["sokrates*"]

[project.scripts]
sokrates-mcp = "sokrates_mcp.main:main"
